Moved BooksForm and BooksList to /containers
